requirements.txt
langchain
langchain_core
langchain_community
langchain-groq
python-dotenv
pypdf
PyPDF2
faiss-cpu
groq
streamlit


